plop

+++ raspberry/autonome3.py
@@ -259,7 +259,7 @@ class MySend(Thread):
             #------------------------------------------------- CALCUL COMMANDES ----------------------------------------------------
             
             if self.enable:
-                cmd_turn = 50 + self.turn*50 | 0x80
+                cmd_turn = 50 + int(self.turn*50) | 0x80
                 #print(cmd_turn)
                 if MySend.differentielD :
                     cmd_mv_droit = (60 - self.move*self.speed_cmd) | 0x80   #marche arrière
@@ -271,7 +271,7 @@ class MySend(Thread):
                     cmd_mv_droit = (50 + self.move*self.speed_cmd) | 0x80
                     cmd_mv_gauche = (50 + self.move*self.speed_cmd) | 0x80
             else:
-                cmd_turn = 50 + self.turn*50 & ~0x80
+                cmd_turn = 50 + int(self.turn*50) & ~0x80
                 cmd_mv_droit = (50 + self.move*self.speed_cmd) & ~0x80
                 cmd_mv_gauche = (50 + self.move*self.speed_cmd) & ~0x80
